requirements.txt
flask
gunicorn==22.0.0
flask-sqlalchemy
flask-migrate
python-dotenv